+++ mitra_api/src/mastodon_api/notifications/views.rs
@@ -13,24 +13,38 @@ use actix_web_httpauth::extractors::bearer::BearerAuth;
 use mitra_activitypub::authority::Authority;
 use mitra_config::Config;
 use mitra_models::{
-    database::{get_database_client, DatabaseConnectionPool},
+    database::{
+        get_database_client,
+        DatabaseConnectionPool,
+    },
+    markers::{
+        queries::{create_or_update_marker, get_marker_opt},
+        types::Timeline,
+    },
     notifications::queries::{
         delete_notifications,
         get_notifications,
     },
 };
 
-use crate::http::get_request_base_url;
-use crate::mastodon_api::{
-    auth::get_current_user,
-    errors::MastodonError,
-    media_server::ClientMediaServer,
-    pagination::{get_last_item, get_paginated_response},
+use crate::{
+    http::{
+        get_request_base_url,
+        JsonOrForm,
+    },
+    mastodon_api::{
+        auth::get_current_user,
+        errors::MastodonError,
+        media_server::ClientMediaServer,
+        pagination::{get_last_item, get_paginated_response},
+    },
 };
+
 use super::types::{
     Notification,
     NotificationQueryParams,
     NotificationPolicy,
+    ReadNotificationsForm,
 };
 
 // https://docs.joinmastodon.org/methods/notifications/#get
@@ -48,6 +62,14 @@ async fn get_notifications_view(
     let base_url = get_request_base_url(connection_info);
     let authority = Authority::from(&config.instance());
     let media_server = ClientMediaServer::new(&config, &base_url);
+    let maybe_marker = get_marker_opt(
+        db_client,
+        current_user.id,
+        Timeline::Notifications,
+    )
+        .await?
+        // Ignore invalid last_read_id values
+        .and_then(|marker| marker.last_read_id.parse::<i32>().ok());
     let notifications: Vec<Notification> = get_notifications(
         db_client,
         current_user.id,
@@ -60,6 +82,7 @@ async fn get_notifications_view(
             &authority,
             &media_server,
             item,
+            maybe_marker,
         ))
         .collect();
     let maybe_last_id = get_last_item(&notifications, &query_params.limit)
@@ -98,6 +121,25 @@ async fn notification_policy_view(
     Ok(HttpResponse::Ok().json(policy))
 }
 
+// https://docs-develop.pleroma.social/backend/development/API/pleroma_api/#apiv1pleromanotificationsread
+#[post("/read")]
+async fn read_notifications_view(
+    auth: BearerAuth,
+    db_pool: web::Data<DatabaseConnectionPool>,
+    form: JsonOrForm<ReadNotificationsForm>,
+) -> Result<HttpResponse, MastodonError> {
+    let db_client = &**get_database_client(&db_pool).await?;
+    let current_user = get_current_user(db_client, auth.token()).await?;
+    let last_read_id = form.into_inner().max_id.to_string();
+    create_or_update_marker(
+        db_client,
+        current_user.id,
+        Timeline::Notifications,
+        &last_read_id,
+    ).await?;
+    Ok(HttpResponse::Ok().finish())
+}
+
 pub fn notification_api_v1_scope() -> Scope {
     web::scope("/v1/notifications")
         .service(get_notifications_view)
@@ -107,4 +149,9 @@ pub fn notification_api_v1_scope() -> Scope {
 pub fn notification_api_v2_scope() -> Scope {
     web::scope("/v2/notifications")
         .service(notification_policy_view)
+}
+
+pub fn notification_api_pleroma_v1_scope() -> Scope {
+    web::scope("/v1/pleroma/notifications")
+        .service(read_notifications_view)
 }